# openapi.yaml
openapi: 3.1.0
info:
  title: ChatGPT Bridge API
  version: 0.1.0
  description: >-
    Controlled access to a Linux development workspace. Paths are always relative
    to the configured workspace root. Shell commands run with the permissions of
    the operating-system user running chatgpt-bridge.
servers:
  - url: https://bridge.example.com
    description: Replace with your public HTTPS bridge URL.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: chatgpt-bridge
        version:
          type: string
          example: 0.1.0
      required: [status, service, version]

    InfoResponse:
      type: object
      properties:
        service:
          type: string
        version:
          type: string
        workspace_root:
          type: string
        default_timeout_ms:
          type: integer
          minimum: 1
        max_timeout_ms:
          type: integer
          minimum: 1
        max_output_bytes:
          type: integer
          minimum: 1
        max_file_bytes:
          type: integer
          minimum: 1
        capabilities:
          type: array
          items:
            type: string
      required:
        - service
        - version
        - workspace_root
        - default_timeout_ms
        - max_timeout_ms
        - max_output_bytes
        - max_file_bytes
        - capabilities

    ExecRequest:
      type: object
      properties:
        command:
          type: string
          description: Shell command executed with /bin/sh -lc.
        cwd:
          type: string
          description: Workspace-relative working directory. Empty or omitted means workspace root.
        timeout_ms:
          type: integer
          minimum: 1
          description: Optional timeout in milliseconds. Must not exceed the server maximum.
      required: [command]

    ExecResponse:
      type: object
      properties:
        success:
          type: boolean
        exit_code:
          type: [integer, "null"]
        stdout:
          type: string
        stderr:
          type: string
        stdout_truncated:
          type: boolean
        stderr_truncated:
          type: boolean
        duration_ms:
          type: integer
          minimum: 0
      required:
        - success
        - exit_code
        - stdout
        - stderr
        - stdout_truncated
        - stderr_truncated
        - duration_ms

    ReadFileRequest:
      type: object
      properties:
        path:
          type: string
          description: Workspace-relative UTF-8 text file path.
      required: [path]

    ReadFileResponse:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        bytes:
          type: integer
          minimum: 0
      required: [path, content, bytes]

    WriteFileRequest:
      type: object
      properties:
        path:
          type: string
          description: Workspace-relative UTF-8 text file path.
        content:
          type: string
        create_parents:
          type: boolean
          default: false
        overwrite:
          type: boolean
          default: false
      required: [path, content]

    WriteFileResponse:
      type: object
      properties:
        path:
          type: string
        bytes:
          type: integer
          minimum: 0
        created:
          type: boolean
      required: [path, bytes, created]

    ListFilesRequest:
      type: object
      properties:
        path:
          type: string
          description: Workspace-relative directory path. Empty or omitted means workspace root.
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 200

    DirectoryEntry:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        kind:
          type: string
          enum: [file, directory, symlink, other]
        size:
          type: [integer, "null"]
          minimum: 0
      required: [name, path, kind, size]

    ListFilesResponse:
      type: object
      properties:
        path:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/DirectoryEntry"
        truncated:
          type: boolean
      required: [path, entries, truncated]

    ChangeSummary:
      type: object
      properties:
        created: { type: integer, minimum: 0 }
        modified: { type: integer, minimum: 0 }
        deleted: { type: integer, minimum: 0 }
      required: [created, modified, deleted]

    BeginChangeRequest:
      type: object
      properties:
        cwd:
          type: string
          description: Workspace-relative project directory to checkpoint. Empty or omitted means workspace root.

    BeginChangeResponse:
      type: object
      properties:
        transaction_id: { type: string }
        scope: { type: string }
      required: [transaction_id, scope]

    FinishChangeRequest:
      type: object
      properties:
        transaction_id: { type: string }
      required: [transaction_id]

    FinishChangeResponse:
      type: object
      properties:
        changed: { type: boolean }
        change_id:
          type: [string, "null"]
          description: Stable checkpoint ID. Null when the transaction produced no filesystem changes.
        summary:
          $ref: "#/components/schemas/ChangeSummary"
      required: [changed, change_id, summary]

    RestoreCheckpointRequest:
      type: object
      properties:
        id:
          type: string
          description: Full checkpoint ID such as chg-A1B2C3 or its unique short suffix A1B2C3.
        force:
          type: boolean
          default: false
          description: Preserve uncheckpointed current changes in a safety checkpoint before restoring.
      required: [id]

    RestoreCheckpointResponse:
      type: object
      properties:
        restored_to: { type: string }
        scope: { type: string }
        safety_checkpoint_id:
          type: [string, "null"]
      required: [restored_to, scope, safety_checkpoint_id]

    CheckpointInfo:
      type: object
      properties:
        id: { type: string }
        scope: { type: string }
        parent:
          type: [string, "null"]
        created_at_ms:
          type: integer
          minimum: 0
        summary:
          $ref: "#/components/schemas/ChangeSummary"
      required: [id, scope, parent, created_at_ms, summary]

security:
  - bearerAuth: []

paths:
  /health:
    get:
      operationId: healthCheck
      summary: Check whether the bridge is running.
      security: []
      responses:
        "200":
          description: Bridge is alive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /v1/info:
    get:
      operationId: getBridgeInfo
      summary: Get workspace and capability information.
      responses:
        "200":
          description: Bridge information.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfoResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/exec:
    post:
      operationId: executeCommand
      x-openai-isConsequential: false
      summary: Run a shell command in the development workspace.
      description: >-
        Use this for builds, tests, Git commands, searches, and other terminal work.
        The cwd must be relative to the configured workspace root.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecRequest"
      responses:
        "200":
          description: Command completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecResponse"
        "400":
          description: Invalid command request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "408":
          description: Command timed out.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/files/read:
    post:
      operationId: readWorkspaceFile
      x-openai-isConsequential: false
      summary: Read a UTF-8 text file from the workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadFileRequest"
      responses:
        "200":
          description: File contents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadFileResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: File not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/files/write:
    post:
      operationId: writeWorkspaceFile
      x-openai-isConsequential: false
      summary: Create or replace a UTF-8 text file in the workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteFileRequest"
      responses:
        "200":
          description: File was written.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WriteFileResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: File exists and overwrite was not enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/files/list:
    post:
      operationId: listWorkspaceDirectory
      x-openai-isConsequential: false
      summary: List a directory in the workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListFilesRequest"
      responses:
        "200":
          description: Directory listing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListFilesResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/changes:
    get:
      operationId: listCheckpoints
      summary: List completed workspace change checkpoints, newest first.
      responses:
        "200":
          description: Visible change checkpoints.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckpointInfo"

  /v1/changes/begin:
    post:
      operationId: beginChange
      x-openai-isConsequential: false
      summary: Start a checkpoint transaction before mutating a project.
      description: >-
        Call this immediately before the first mutating command or file write for a user request.
        The snapshot store is external to the project and never snapshots .git, .hg, or .svn metadata.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeginChangeRequest"
      responses:
        "200":
          description: Transaction started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeginChangeResponse"

  /v1/changes/finish:
    post:
      operationId: finishChange
      x-openai-isConsequential: false
      summary: Finish a checkpoint transaction after edits and tests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinishChangeRequest"
      responses:
        "200":
          description: Returns a change ID only when filesystem state changed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinishChangeResponse"

  /v1/changes/restore:
    post:
      operationId: restoreCheckpoint
      x-openai-isConsequential: false
      summary: Restore a project to the state after a selected change checkpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreCheckpointRequest"
      responses:
        "200":
          description: Workspace restored to the selected checkpoint.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreCheckpointResponse"
        "409":
          description: Current workspace contains uncheckpointed changes and force was not enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/changes/undo:
    post:
      operationId: undoCheckpoint
      x-openai-isConsequential: false
      summary: Restore the state immediately before a selected change checkpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreCheckpointRequest"
      responses:
        "200":
          description: Workspace restored to the checkpoint parent state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreCheckpointResponse"